+++ onprem/core.py
@@ -265,16 +265,19 @@ class LLM:
         chunk_overlap: int = 50, # character overlap between chunks in `langchain.text_splitter.RecursiveCharacterTextSplitter`
         ignore_fn:Optional[Callable] = None, # callable that accepts the file path and returns True for ignored files
         pdf_use_unstructured:bool=False, # If True, use unstructured for PDF extraction
+        **kwargs, # Extra kwargs fed to `langchain_community.document_loaders.pdf.UnstructuredPDFLoader` when pdf_use_unstructured is True
     ):
         """
         Ingests all documents in `source_folder` into vector database.
         Previously-ingested documents are ignored.
+        Extra kwargs fed directly to `langchain_community.document_loaders.pdf.UnstructuredPDFLoader` when pdf_use_unstructured is True
         """
         ingester = self.load_ingester()
         return ingester.ingest(
             source_directory,
             chunk_size=chunk_size, chunk_overlap=chunk_overlap, ignore_fn=ignore_fn,
-            pdf_use_unstructured=pdf_use_unstructured
+            pdf_use_unstructured=pdf_use_unstructured,
+            **kwargs
         )
 
     def check_model(self):